requirements.txt
discord.py==1.3.4
aiohttp
numpy
pillow
websockets
git+https://www.github.com/hbldh/hitherdither
fuzzywuzzy
python-Levenshtein
psutil
git+https://github.com/Rapptz/discord-ext-menus
sqlalchemy
psycopg2
